# Remove `find_or_raise_by_id` in `_set_trial_value_without_commit`