## Motivation
`RDBStorage` uses `find_or_raise_by_id` to retrieve a `TrialModel` in `_set_trial_value_without_commit`, but it can be directly obtained from callers.


## Description of the changes
By passing the trial model as an argument, this PR removes the need for `find_or_raise_by_id` in `_set_trial_value_without_commit`.

## Benchmark

```python
import optuna

def objective(trial: optuna.Trial) -> float:
    x = trial.suggest_float("x", -100, 100)
    y = trial.suggest_int("y", -100, 100)
    return x**2 + y**2

sampler = optuna.samplers.TPESampler(seed=42, multivariate=True, constant_liar=True)
study = optuna.create_study(sampler=sampler, storage="sqlite:///tmp.db")
study.optimize(objective, n_trials=1000)

```

| master | PR |
| - | - |
| 10.30s | 10.02s |